requirements.txt
psycopg2-binary==2.7.5
vk-api==10.1
requests==2.18.4
raven>=3
